requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-dotenv
requests
apscheduler
streamlit
pytest